CodeEnvironment/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-default-group==1.2.4
cloup==2.1.2
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
filelock==3.13.4
fire==0.6.0
flake8==7.0.0
Flask==3.0.3
Flask-Cors==4.0.0
fonttools==4.51.0
glcontext==2.5.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isosurfaces==0.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
kiwisolver==1.4.5
manim==0.18.0.post0
ManimPango==0.5.0
mapbox-earcut==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mccabe==0.7.0
mdurl==0.1.2
moderngl==5.10.0
moderngl-window==2.4.4
multipledispatch==1.0.0
networkx==3.2.1
numpy==1.26.4
openai==1.23.2
packaging==24.0
pandas==2.2.2
Pillow==9.5.0
pycairo==1.26.0
pycodestyle==2.11.1
pydantic==2.7.0
pydantic_core==2.18.1
pydub==0.25.1
pyflakes==3.2.0
pyglet==2.1.dev2
Pygments==2.17.2
pyobjc-core==10.2
pyobjc-framework-Cocoa==10.2
pyparsing==3.1.2
pyrr==0.10.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
scipy==1.13.0
screeninfo==0.8.1
setuptools==69.5.1
six==1.16.0
skia-pathops==0.8.0.post1
sniffio==1.3.1
srt==3.5.3
svgelements==1.9.6
termcolor==2.4.0
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
watchdog==3.0.0
wcwidth==0.2.13
Werkzeug==3.0.2
